pyproject.toml
[project]
name = "flippers"
version = "0.1.0"
readme = "README.md"
description = "A burger joint simulator"
authors = [{ name = "Alex Cross", email = "AlexJCross90@gmail.com" }]
requires-python = ">=3.13"
dependencies = ["hydra-core>=1.3.2", "pydantic>=2.11.9"]

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[project.scripts]
flippers = "flippers.cli:main"

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"

[tool.ruff]
line-length = 110
target-version = "py313"
